requirements.txt
aiocache==0.12.2
aiohttp==3.9.5
aioredis==2.0.1
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
async-timeout==4.0.3
attrs==23.2.0
black==24.4.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Deprecated==1.2.14
dill==0.3.8
discord.py==2.4.0
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
fakeredis==2.23.3
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-discord==0.2.6
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
invoke==2.2.0
isort==5.13.2
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mongomock==4.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==2.0.1
oauthlib==3.2.2
openai==1.35.13
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyGithub==2.3.0
Pygments==2.18.0
PyJWT==2.8.0
pylint==3.2.6
pymongo==4.8.0
PyNaCl==1.5.0
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
redispy==3.0.0
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
schedule==1.2.2
sentinels==1.0.0
shellingham==1.5.4
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
tenacity==8.5.0
tiktoken==0.7.0
tomli==2.0.1
tomlkit==0.13.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
vulture==2.11
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
